+++ src/lib/validators/account-credentials-validator.ts
@@ -3,7 +3,7 @@ import { z } from "zod";
 export const AuthRegisterCredentialsValidator = z.object({
     username: z.string().min(3, "Name is required"),
     email: z.string().email(),
-    password: z.string().min(8)
+    password: z.string().min(8),
 })
 
 export const AuthCredentialsValidator = z.object({